+++ tests/conftest.py
@@ -537,7 +537,14 @@ def login(request):
         "first_name": "Rudá",
         "email": "rudazz@gmail.com",
         "last_name": "Filgueiras",
-        "groups": ["g:briefy_qa", "g:briefy_pm", "g:briefy_bizdev", "g:briefy_scout", "g:briefy"]
+        "groups": [
+            "g:briefy_qa",
+            "g:briefy_pm",
+            "g:briefy_bizdev",
+            "g:briefy_scout",
+            "g:briefy_finance",
+            "g:briefy"
+        ]
     }
     policy = JWTAuthenticationPolicy(private_key=JWT_SECRET,
                                      expiration=int(JWT_EXPIRATION))